Extract `handle_inspect` from the main module
